requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
black==24.1.1
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.9
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
pytest==8.0.0
pytest-celery==0.0.0
pytest-django==4.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
